backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.21
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
mysqlclient==2.1.1
redis==5.0.1
celery==5.3.1
stripe==6.6.0
requests==2.31.0
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
python-dateutil==2.8.2
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0 